LayoutTests/accessibility/mac/treeitem-row-delegation-expected.txt

This tests that a treeitem was a group that uses aria-owns will report its disclosed rows correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS treeitem2.isEqual(accessibilityController.accessibleElementById('group1-item1')) is true
PASS successfullyParsed is true

TEST COMPLETE
Jungle Animals
Lions
